property sheet: Support auto complete in text fields
This CL adds support for auto completing resource references.
The list of possible references matches what is shown in the
ChooseResourceDialog (which is what is shown when you click
on "Browse.." next to the text field), but will be tweaked
in a future CL to reuse the logic used by the XML editor to
determine possible completions.

Change-Id: Iebceec4fece83225cc27557dcd4f578b6e092583