Nele: Include the preview form in the active tool editors.
The preview form should be able to interact with a floating tool palette.
This did not work because the preview was ignored as a possible editor.
After this fix there are still a couple of problems:
1) When a floating palette is opened the palette for the preview is
   disposed twice => NullPointerException.
2) The palette for the preview has an option to be placed on the right,
   but that is actually not implemented, so eht option should be removed.

Change-Id: I9ca2f34f78c7536e1c3282fdb3220393b741f5db
(cherry picked from commit 2ba3d0e7724e79df7a1799fdb4f1283c19b4d5f1)