Add qualifiers support to the model
Adds all the available attributes to EditedStyleItem. One of the values
will be marked as selected by the current configuration.
This is part of adding qualifiers support to the theme editor. This
change doesn't do any UI changes and everything should work the same
from the UI point of view.

Change-Id: I23669f76d78dd0157a6eae00cd28268ae01aca37