Fix flaky UI test for translations editor
The UI test for translations editor had a condition that would
make it wait until the table showing the translations showed up.
However, that is not sufficient since the table itself is
populated asynchronously. So we now wait for the loading panel
that is overlaid on top of the table to indicate that it has
finished loading before proceeding with checking the content
of the table.

Change-Id: I4115d27174c7a426bdd160edaeac22895cf1930c