Nele: Fix problem with hierarchy updates in NlModel.
When an XML files is parsed e.g. after a user changes the XML by hand in
the "Text" pane of the designer, some of the components end up having
the TagName property set to something different from Tag.Name.
It looks like the XmlTag instances have been reallocated to the same
spot in memory and the hash in NlModel is still pointing to an older
version. In that case the TagName in NlComponent end up with the
wrong value which causes bad behaviour in the structure panel and
the designer.
This is a bandaid - we should reevaluate the use of the hash in
NlModel.

Change-Id: I65ca3af2e7bc8536af440628e2353016e5684b50