Fix theme attribute resolution
This CL changes the attribute resolution. Before, the attribute
resolution would be done from child to parent and this CL reverses that
and also generates a value for every configuration where the child theme
is defined.

This addresses the problem where a child theme might be defined for
"v21" and "default" but one attribute was only defined in "default". When
resolving v21, the previous code would select the attribute defined in
the child "default" but the correct behaviour was to select a value from
the parent.

Change-Id: Ic6d1254cecb10fb8f6acc7517b4c79ca0078e109