Logcat no longer discards last dead client
Previously, a user debugging a process in logcat would have their
profiling views forcefully closed if their process crashed or was
closed.

Now, if a client is removed that was currently selected, we keep it
in the list but render its title as [DEAD].

This CL also disables all profiling buttons when the target is a
dead client.

Related issue: http://b.android.com/161145

Change-Id: I0f7f0b7004f771116b6c07991b3bed55b9dbee64