Nele: Add dependency checks for palette.
The components on the palette that belong to a library
that is not included in the current module will no longer
be able to be dragged onto the designer.
If the user clicks on the component on the palette, a dialog
prompts the user if the library should be added to the module.
Changes to the known list of missing libraries for the palette
is detected from project save operations.

Change-Id: I2b1a1093b9214c5c359192fb854e81f1bc28485d