Fix simple mode sorting and UI tests
Before this CL sorting of the attributes was done in the model where
attributes are sorted within their categories (Style, Drawable, Color,
etc). When filtering is applied in simple mode, the sorting is wrong
since the categories are removed.
This fixes the sorting by applying the sorting at the JTable label.

+ This also fixes the UI test since it wasn't expecting the namespace.
+ Add test for AttributeGrouper.

Change-Id: Idd8c8e9f876f872faf86364a80cceb60bb9df9c4