Draft a generic AndroidModel interface.
This interface is meant to replace current usages of IdeaAndroidProject
in non-Gradle-specific code. In particular, it abstracts away the
details of product flavors/build types/variants and hides the
Gradle-specific concept of the "selected variant".

Note that the adoption of this interface will force the consolidation of
the AndroidFacet#get**SourceProvider methods. Based on current usage I
see only a couple places where this could be tricky; we are
investigating those now.

Note also that this interface pulls down some methods from
AndroidModuleInfo (which is currently used in part as an indirection
layer between Gradle and non-Gradle Android modules). With a bit of
work, we might be able to collapse these layers of indirection into one.

Change-Id: I065807e05febb22bf1dc6687a059981a68d2fe55