Semi-fixing GTK+ issue in sdk manager
Previously when using GTK+ theme the SDK manager would paint
the selected package name as white text on a white background.
Now it paints it correctly, except that the part of the "name"
column of the selected row that isn't actually the label itself
(that is, the blank part on the sides of the checkbox and text)
are painted as a different color from the rest of the row. It
seems this is a limitation of TreeTable and GTK+, as it is seen
elsewhere in IJ and Studio.
Fixes https://code.google.com/p/android/issues/detail?id=181490

Change-Id: I3dd82a21b25575688035c3cec499d3a654ce931d