Fix IndexNotReadyException in layout editor
Layout editor used to display an error with IndexNotReadyException
in dumb mode, steps to reproduce:
  1. Enter dumb mode
  2. Replace layout file contents in XML editor with incorrect XML,
     I've used one line "Test"
  3. Open "Design" view of activity, error pane should show up

The reason is that method AndroidFacet#fillMap uses PsiFacade for
finding a class by name inside runReadAction, which can throw
IndexNotReadyException. This CL just wraps call to facade.findClass
in try-catch block. Another possible way to handle that might be to
use alternative ways to resolve classes, but fillMap already handles
possibility that class might be unavailable and UI shows a reasonable
error message in that case.

Change-Id: I308612dc262fd6feee71d9ebb990aabc9e7f8b38