Add appCompatActivity to the wizard data.
The key appCompatActivity is set to true if we have
a detected appCompat as a library (or this is a new
project) and the build version is 22 or higher.
These are the conditions for using AppCompatActivity
as the base class for a new activity.
By adding this value to the data we can use this value
for determining the thumb nail shown in the new
activity & new project wizard.

Change-Id: Idee0cc41d5b0d821001fb5fab9c220e934881df5