Fixing addon-dependant formfactors in npw
API was changed out from under us by
https://android-review.googlesource.com/147786. Updating
to use new API.
Also doing a lot of refactoring. Notablably, previously if
a platform was unavailable a download link would show, and after
clicking and downloading the form factor would become selectable.
This is dumb, though, since the next step already installs needed
packages. So made this step just add the necessary addon to the
requested packages instead of forcing it to be downloaded first.

Change-Id: I308cb9c6cc59c8841fa189852cc5347e367577e3