Stop the samplers if the application goes away
Fix for 180648

I moved lots of code from the individual sampling monitors up to the
base classes so I could fix the issues in just one place.
I made it so that stop does not discard the data, but leaves it visible
underneat the disabled message just like pause does.
I call stop when the client becomes invalid (which is what happens when
the process the client is connected to dies)

Change-Id: Ieda05cbf814ab97ca28ca6bcc788e5a6295c2a72