Check for valid JDK before project sync.
In AS, we use the same JDK for both IDEA and Gradle. When the jdk is
not valid during the sync, Gradle is throwing an error with a quick
fix which is not applicable in AS. To avoid it, now checking for the
JDK validity before sync and reporting an actionalble error to the
user.

Fix for https://code.google.com/p/android/issues/detail?id=172714

Change-Id: Iff463738fd943b072e55af57c0a55130388e4fbe
(cherry picked from commit 8d0f43283cdb6fe62ed8a9a8cab6149363e16f80)