Add new "values" constraint for template parameters
Currently, when creating a new "XML values file", we do not check if the
destination file already exists. This can destroy the contents of files
like strings.xml or dimens.xml without warning.
This CL adds a new "values" constraint that, when combined with "unique"
allows to specify that we do not want to override files.

Bug: http://b.android.com/164628
Change-Id: I41cdd0b47ff2217de1b3b38ac2c0c80204c7c2de