Refactor AndroidUtils.executeCommandOnDevice
I was reading the code while working on logcat issues.
Originally I was just going to fix the fact that it was using a deprecated function (needed to use the TimeUnits version of executeShellCommand).
I decided to merge the two calls, but I found the code quite hard to read, so I refactored a little on my way through.

Change-Id: I38837f795c8c2b396aca5f65fb035c51d78529b9