Revert "navigator: Remove AndroidTreeUpdater and remove some unused imports"
This reverts parts of commit 7e1707c14ce523807c77f31e939de33a2e126659.

I originally reverted this code thinking that it wasn't necessary,
but it turns out that this allows a proper refresh of the tree
hierarchy. Also added a few null checks.

Fixes a couple of bugs: 183979, 179463, etc.

Change-Id: Ic3bd056c382981c6215397f3d711139b87eadf05