Remove intermediate allocations done by StackTraceExpander
StackTraceExpander returned a list of strings every time you processed
a line of console output. This was inefficient when the processed
result was just a single line of text, which had to get wrapped in a
temporary list object every time; worse when you consider this case
was extremely common, probably 99%+ of the time.

Now StackTraceExpander maintains its own list of processed text and
exposes it through a getter that you're expected to check after
calling process().

Change-Id: If89aa3c53886a8bbe5451803158c55233efc781b