Prefer support library for @Nullable
If a class has an unresolved reference to @Nullable, the first
quickfix offered by the plugin is "add annotations.jar to classpath,
which is used to convert the reference to the org.jetbrans.annotations
version of @Nullable. In Android projects, suppress this behavior
such that a later quickfix instead suggests adding the support
library annotations as a Gradle dependency.

Also update the fix message to match the format used by other
add Gradle dependency quickfixes.

Change-Id: I559ab9280742e4df64986ca72d80510b06daf372