93265: Graphical layout editor creating duplicate IDs
The ID manager is responsible for assigning unique
id's, across layouts. It uses the resource repository
to do so.

However, in a RelativeLayout, there are many separate
calls to assign id's in rapid succession, and when
this happens the resource repository hasn't been
updated yet, which can lead to the same id being
assigned more than once. That's Bad(tm).

To avoid this, we also scan the local component hierarchy
and include those id's in the set of known id's.

Change-Id: I172408d5f6d6fdc6493e2412fe4365df1834a265
(cherry picked from commit 4a99eeab9d3e5041152b50c17742cf4f9ddf3554)