Use immutable sets in value resources index.
Deep in the guts of the IJ indexing ValueContainerImpl, hashcodes are
very frequently computed for the stored values. HashSet's hashCode
implementation is linear in the size of the set and *not* memoized.
ImmutableSet hashCode is constant time.

By storing immutable sets of MyResourceInfo instead of mutable HashSets,
we see 18% performance improvements on calls to
AndroidResourceUtil#findResourceFields (when PSI cache is cold), e.g.
when (on the UI thread) we respond to a find-references call.

Change-Id: I292703d35441319b064d251b2587932b94f830d0