Introduced notion of uncommitted state to build.gradle models.
This CL introduces GradleDslModel, which replaces GradleDslElement. The
main differences are:
- GradleDslModel is an abstract class (instead of being an interface,
  as GradleDslElement)
- GradleDslModel has the notion of "uncommitted state", which is critical
  for using the new build.gradle parser in the "Project Structure" dialog.

The "Project Structure" dialog displays a modifiable representation of
the project's build.gradle files. It also works by making modifications
to that model without committing those changes. The user can either
"apply" any changes made (e.g. changed the version of a dependency) or
"cancel" or changes made.

With this CL, models created by parsing build.gradle files now have the
new methods:
- public boolean isModified()
- public void applyChanges()
- protected void setModified(boolean)

In addition, a GradleDslModel keeps a reference to its parent (another
GradleDslModel, nullable). When a model is modified, it notifies its
parent, and the parent notifies its own parent, and so on. This way
GradleBuildModel (the top model) will now if any of its children has
been modified.

This CL only makes changes to the external and module dependency parsers.
The Java and Android DSL parsers need to be updated. Also, the method
DependenciesModel#addExternalDependency needs to be updated to this new
way of handling changes.

Change-Id: Ic3ebadbacc7e830e10a6504cd46cd97ac61b38ec