Problem getting resource dir in vector asset wizard
Previously if you selected a file in a module not part of a
SourceProvider (e.g. a module gradle file), and did "new vector
asset" there would be a problem due to no source provider being
found. Updated to check for a source provider for the containing
directory (if a file is selected) rather than for the file itself,
falling back to getting the default resource dir for the module
as a last resort.
Also, we were showing the project-level pseudo-module in the module
chooser. Filtering it out.
Fixes https://code.google.com/p/android/issues/detail?id=186798

Change-Id: I0d2a8b408c97683fb3c20f350f57b5998b6ade13