Pass in created project to importNewlyCreatedProject
NewProjectWizardDynamic first creates a project in doFinish().
It then calls importNewlyCreatedProject in its runFinish() method.
However, since it doesn't pass in the project that it already
created, this results in the import action creating yet another
project. So we end up with 2 project objects in memory, while
only one of them is properly initialized. When the IDE is closed,
one of them doesn't get disposed off properly resulting in errors.

Change-Id: Ie91f23c7b64b7079228a6af3062c96f09e6abeb1