Use ImageUtils.scale instead of Image.getScaledInstance.
Fix for issue: 187323.
The images used in the gallery for "New Module" currently look bad.
They have smudge at the bottom of most images.
This can be fixed by scaling the image down in steps like it is done
in ImageUtils.scale and uploading some new images that are more
scaling resilient (from Gerard).

Change-Id: I32ca477b3d9855db6ec9908dd3f38dfcdd1a40a4