Limit the size of the drawable preview render
Currently, our code will call layoutlib and use the intrinsic size of
the icon to render the drawable. This is, in most cases, a wast of space
and time since the icon will be scaled down to the size of the swatch
icon. In the worst case scenario, if the drawable doesn't have an
intrinsic size, a device size will be use to render.
This CL speeds up significantly the paint of the attributes table and
reduces the GC trashing.

Change-Id: I44aadac5c8d115c3e27408df6fa04ab9dad6664f