Extract an emulator launch options class.
Rather than keeping all the state and logic for computing the emulator
launch command line in AndroidRunConfigurationBase, extract a separate
class for keeping the settings and computing the command line.

The fields have the same names and are still serialized via the
DefaultJDOMExternalizer, so the serialization in compatible with older
versions.

Change-Id: Ia8a2b292243127d5bee7ae58a2c79ecc7eba8da8