Remove obsolete fast deploy code
This CL removes a lot of early prototyping code related
to fast deploy: non-Gradle hot swapping of resources
and direct computation of .class and .dex files, the
earlier "fast deploy mode" action which pushed changes
across on save, etc. Also some cleanup-renames.

It also adds basic support for checking whether the app
is running or not, and whether the Gradle plugin is
compatible or not; this will be used by the Run configuration
code to make the Run button overload Restart/Patch.

Change-Id: I5916a5c2979a340662ece6ebfc91bec194fc97a0