Allow servers to restart
These changes tie the shutdown of gapir to the shutdown of gapis, which
happens when the last active editor goes away. This also means that
gapir is torn down correctly when studio shuts down.
Also is the ability to connect to a new gapis and gapir if you open a
new editor again after closing the last active one before.
If you kill gapis and gapir by hand (using killall or something) then
active editors will not attempt to reconnect, but new editors will run
up a brand new server and keep going.

Change-Id: I652855432a0135ebf8ab573634199d5e9fc0fd93