Check whether app is in the background and if so, restart it
If an app is placed in the background (for example switching back
to a different app, or starting a new one via a notification) then
it may still respond to pings from the IDE, which makes the IDE
just do an incremental push instead. However, the user expectation
on pressing Run is to get the app fronted and updated again (as
would happen for a non-incremental Run), so we now look for the
foreground/background state in the reply from the server ping,
and use that to only do an incremental run if the app is running
*and* is in the foreground. In a followup CL, we can use the
background state to issue an adb activity manager command to
get the app to either front or restart.

Change-Id: Ic71540c636cf9fa7e6d23c51fc2e88bf0cd500f1