Remove redundant setTargetPackage call from debug runner
AndroidDebugRunner was explicitly setting the target package name
under a very specific circumstance: it is a test debug run, and this
is a non Gradle project. However, once the debug runner was done,
the same field was overwritten by AndroidRunningState anyway. So
it seems like a pointless operation. This CL removes that from
AndroidDebugRunner.

Change-Id: I71c2a48caafd2a5f86f79067355c6e900ecebe3d