Cleanup application run parameters
ApplicationRunParameters is a configuration specific editor. This gets
plugged into the run configuration editor for an android application
launch. It provides the UI for two major choices: what to deploy, and
how to launch the application.

This CL changes the UI and the implementation for it. In both cases, the
earlier UI used radio buttons to show all the options. This was fine
earlier, but we need to include more options for launch (such as
launching a deeplink). Hence both the options are now implemented as
combo boxes. In both cases, selecting a particular deploy or launch
option in the combo will update the form to show other options that are
specific to the selection (Earlier all these sub-options were always
visible, resulting in a cluttered UI).

The implementation looks longer, but is fairly straightforward: Each of
the options for launch is implemented as a separate class. This allows
each class to maintain the state corresponding to itself separately
instead of all these options being stored inside the main class. Each
option also provides the custom UI that is shown when that option is
selected.

TODO:
 - pm options need to be hooked in to pm install

Change-Id: I25a73cd193a5b8f1db724c2a6b3dcaa8b128f8ec