Add Instant Run diagnostics and fix link
If you open the Instant Run options panel, it now shows you
custom labels telling you that the project needs to be built
with Gradle (if it's not), or that you need a more recent
version of the Gradle plugin (if needed). Clicking on the
link to update the version will update the plugin and sync.

Change-Id: If1ebee04f79049aee20ddc5cb8bc1dba3df662a1