Hook up the pm install options from the UI to the actual launch
In a previous CL, a UI option was added to the launch configuration
editor to allow users to specify custom options to "pm install" command.
This CL hooks up that user input to the actual step that does the
installation.

Fixes https://code.google.com/p/android/issues/detail?id=182281

Change-Id: I07b9e3d70398c1de4d22be6a4f6e5b14ddbe98f0