Removing files on cancel and adding write-to-temp-file.
This CL adds two features/fixes:

1) When calling CaptureService.cancelCaptureFile(), the underlying
file is deleted. If the user encounters an error but wishes to keep
the already-appended-to file (e.g. streaming files with no end), then
the user can just call finalize.

2) Adds a boolean flag to CaptureService.startCaptureFile to write to
a temporary (filename + ".temp") file that won't get picked up by the
CaptureService. This is useful for capture formats that need to be
complete before they can be opened. I.e. the user can't click on the
file while it's still being written to.

Change-Id: I6fe941a8f986a62cfeec8c7cabc635d967161e78