Add actions for activity restart and instant run, change defaults
This CL adds 3 new actions for instant run:
- Restart Activity
- Instant Run (without activity restart)
- Instant Run (with activity restart)

Restart Activity is available in the Debug menu; the other two are
available via Find Action and can be assigned to shortcuts.

Furthermore, the "Restart Activity on Code Changes" option has been
changed to false by default, and instead, when the run action only
results in a hot swap, a toolwindow caption tied to the Run window
appears explaining that the code has changed without an activity
restart, and shows the keyboard shortcut for optionally restarting the
activity (as well as pointing to options to change the default, or to
dismiss the balloon for now or for the rest of the session.)

There are a number of changes to the message protocol too, to make it
more robust (e.g. when an updated IDE talks to an already running
older version of the runtime library). This is described in more
detail in the runtime library CL, but there's a fair bit of code
cleanup in the IDE side of the communication such that the setup and
shutdown is done in a single place.

Change-Id: Iac68fca7cf21afd6aa15fcbb135bad6e12eaa0a0