Perform faster incremental builds when only Java files have changed
The Gradle plugin has an option for the incremental compilation
property to specify that only Java files were changed since the last
build, which lets Gradle skip a bunch of processing stages.

The Fast Deploy manager now tracks file changes (while instant run is
enabled, and while we haven't seen any non-Java changes to the app
module since the last build). If between two builds it doesn't see a
non-Java change, it passes the Java-only flag to Gradle which makes
the build faster. If it does encounter a non-Java change (typically
XML changes) it stops listening for file changes until the next build.

(There is a similar mechanism for resource files, so we're actually
tracking edits to both Java and resource files, and supplying separate
flags.)

Change-Id: Ie7b91e3e0980a5df07d7b7a3b165e111c0569634