Update device picker UX
This CL updates the device picker to match the UX mocks. In particular,
devices are now grouped into two sections: connected devices and
available devices. If an incompatible target is selected, then the
compatibility error is shown as a dialog validation error. Due to the
presence of the separators and info panels inside the list, a special
key listener is installed that allows for a natural keyboard navigation.

A number of features available in the old device chooser have now been
ported over. These include the ability to select multiple devices,
double click listener, "do not ask again" option and more validation.
The device list now also includes speed search capability.

Change-Id: If3bd9190ea8df4a79d92d15dd73f7f39312f2948