Refactor DeployTargets
This CL splits the DeployTarget interface into two parts: a
DeployTargetProvider, and a DeployTarget.

At a high level, a DeployTarget is simply a wrapper on top of either a
custom run profile state implementation, or a list of local devices.
The DeployTargetProvider can provide these DeployTargets either right
after they are configured, or after a runtime prompt to the user.

Change-Id: I61377309e3270efa44fd476bffab2e81de793929