Adding support to resolve references in lists and maps.
Created a common base class called GradleDslExpression for
GradleDslLiteral, GradleDslReference and GradleDslMethodCall
which all are represent the subclasses of GrExpression.

Changed GradleDslListeralList and GradleDslLiteralMap to
GradleDslExpressionList and GradleDslExpressionMap, so that
they can hold the reference and method elements along with
the literal elements.

With this change, variable resolution works fully with in a single
file, when the variables are used wither as direct references or
injections in the string literals.

While at it, also fixed the GradleDslMethodCall to add support to
remove either individual arguments or removing the method call
altogehter.

Change-Id: If82b7142d0da3cb42e49ad8147cf9a555c087581