Synchronize build id even when there are no changes
If the user performs a build, and there are no changes,
we need to copy the current build-id from gradle over to
the device. Without this, Gradle will update its build id,
but the device continues to be tied to the previous build
id, and on the next build the IDE notices the build id's
differ and it triggers a full build.

Change-Id: I07e0f100cf945981d034b9602b74eb049d6ebc6b