Move "resolve parameter" logic to templates package
A class for resolving a template parameter's value (and its custom
exception) was in the npw package although it could safely be moved
to the template package, as it has no dependencies on npw at all.
This seems to be a cleaner place for it, anyway.

Moved the classes, renamed them to shorter versions, tightened up the
API (from constructor+function call to a single static call), and
updated all locations affected by the refactoring.

Change-Id: Ic5462c46d1d669d17db05012b56542448a785efd