ProportionalLayout supports spanning columns
ProportionalLayout is useful for dividing a space up into a grid, but
occasionally there are rows which need to break the grid (for
example, separators, or maybe an image that needs to fill 2 out of
3 spots).

Now, when adding a component, you can specify it lives across
mutiple cells. This removes it from all layout calcuations, but
will size it appropriately based on existing calculations.

Change-Id: I34ac6a94c2cfc2d569e76bc32828a44913a9b5ee