Fix ResourceFolderRepository tests
Seems that after IJ15 merge, some modification events happen in a
different order or for a different reason.
For example, removing a full tag causes a childrenModified in the parent
with the empty spaces that existed before and after the tag that was
removed.

This CL ignores empty space modifications. The change also seems to
reduce the number of cases where a full scan is triggered in some other
test cases.

Change-Id: I5de263f2be9f23196c44e8d31e8b4a544f89591d