Get instant run working on API 15
The recent code to enforce build id's between the device and
Gradle relied on using adb commands to read and write files on
the device, and it can't just use the obvious push and pull
commands to do this because it needs to write to the private
app data directory, even on devices without adb root.

Turns out not all commands are availble everywhere (cp is missing
on API 15 devices, echo is missing on (some) API 19 devices.

This CL tweaks the commands some more to hopefully get this
working everywhere.

Change-Id: I0e3ff8466a577f23e547dc34d8aca09116e397e6