DeviceEditWizard migration to new code.
Changes to relevant and new classes
needed to convert DeviceEditWizard usages to
the new ModelWizard classes.

Affects-docs+2:
Adds a "round" checkbox to the UI so the user can configure
whether their Wear form factor is round
or not (other form factors can't change this).
When the "round" checkbox is toggled, the device preview
updates between a round and rectangular display.

Change-Id: Iaf8cd87633d9aaea570b7d56b4ab8a79bbd9b6d6