Change when the terminate existing session is displayed
We currently show the "do you want to terminate an existing session"
only if we determine that the executor has changed (run -> debug, or
debug -> run). When using the same executor, we used to terminate the
session without the dialog.

This CL changes the behavior such that the dialog is always displayed as
long as it is not a instant run.

Fixes b/25769401

Change-Id: I4d9a2c817bae2bcd06b540932983acb04821c3f7
(cherry picked from commit 88d42dc2ae9368eed9c87b53a9d3dc8a188cc7b0)