Android XML subsystem refactoring
This CL is a step forward moving instanceof checks of
AndroidDomExtender to recently introduced @Styleable annotation. For
making it work with root tags some new interfaces and corresponding
dom file descriptions had to be created, and during that process a
bunch of boilerplate was removed from dom file descriptions by
introducing an AbstractSingleRootFileDescription.

Test for <animation-list> tag in anim/ folder has been removed because
it's incorrect and according to documentaiton <animation-list> XMLs
should reside in drawable/ folder. AndroidDrawableResourcesDomTest
already has a bunch of unit tests for it, so no new unit test were
introduced to replace removed one.

Change-Id: Ib8b5fecbe5868755de23e809a7b6b6feb6db92a0