[UITests] Change importProject to not require synchronous refresh.
We already do a full synchronous refresh in setUp(), running inside
a write action on the EDT, so at least in theory we don't need
importProject to again block on refresh. The problem with importProject
is that it runs on the main thread, so if it ends up in
VfsImplUtil#refreshAndFindFileByPath, it will block until the EDT can
process the refresh queue.

I suspect one common case when the EDT falls behind in refreshing
the VFS state, and when importProject on the main thread hangs forever,
is when a previous failing test left a modal dialog open (because
the EDT is stuck in its inner event loop, see thread dumps in attached
bug).

Bug: http://b.android.com/195103
Change-Id: I544c110ee23a0bfaccbee01cc2cf71120692712f