Invalidate resource val cache after adding styles/array item
The current childAdded() logic handles <item ...> before the
clause "// See if you just added a new item inside a <style>
or <array> or <declare-styleable> etc". If the item doesn't
have a name or type, then the current <item> logic
just returns.

Instead it should fall through and check for the <style>,
<array>, etc. case, and if that fails rescan.

Array items don't have names or types.
Style items have a name, but no type.
Plural items don't have a name (only quantity) or type.

Add some tests for the above, as well as for adding
and removing <item></item> whose parent is not an
<array>, <style>, etc.

Change-Id: I4ef9bd3ff1711cff19f1de8963c2a2ba4d51f640