Implements layoutlib callback getXmlFileParser method
That new method means that layoutlib can now get psi information
for state lists, so it is now unnecessary to save state lists to disk
after they have been modified in the state list picker.
Since adoption of the latest version of layoutlib may be slow,
we leave the code that saves the state lists to disk for the moment.

Bug: http://b.android.com/183767
Change-Id: I81c9891f1381b9f8f67bb072f315b5e810c8e67f