Ensure left-over ResourceType.ID items get removed on deletion.
Previously, the PsiResourceItem was created but not hooked up
to a PsiResourceFile as the source (only added to myItems).
So, when removeFile tries to purge items, it skips these
pending items because their source doesn't match the deleted
file.

Came across this while experimenting with lazy file -> psi
conversion, which needs to be careful about purging all
items from a ResourceFile after a change.

Change-Id: I4a3f0e9c3cd3133d423530d681b03bf28303dcb7