Integrate instant run in regular launch flow
Hotswap (and parts of cold swap) were considered as an entirely
different launch flow and used to bypass the normal run configuration
flow: When the user presses "Run", we make a determination about whether
this is can be hotswapped and if so, we used to call PatchDeployState.

This led to a bunch of issues, the major one being how cold swap was
handled. (Cold swap is very similar to a normal launch flow with the
major difference being how the app is installed). The situation also did
not scale with further refinements in Gradle: there are now plans to do
dex sharding or split apk install in lieu of cold swap depending on
various conditions.

This CL removes the PatchDeployState, and moves Hotswap to be just
another "launch task". Once the build is done, the "LaunchTasksProvider"
examines the output of the build and attempts to determine what steps
need to be taken. Currently, it only supports two flows:
 - If it detects that the change can be hotswapped, then it returns just
   a single task that does the hotswap.
 - Otherwise, it returns a list of tasks corresponding to the actions
   necessary for the run configuration.
Subsequent CLs will tweak this class to implement more methods, e.g we
might do the full flow, but customize just how the app is deployed.

Change-Id: I3bbd508e4126aec4406fb4b8f9101782157bcfef