Don't show incorrect message about the cause of a build
When the IDE is launched for the first time, we don't have any cache of
what was installed on the device. As a result, we decide that a rebuild
is needed because our manifest/resource timestamps don't match the
current state.

This CL makes displaying the cause of the build conditional on whether
the cache had any data.

Change-Id: I3318f985b38c0313bf30eb5f34de6078c330dd7d