Perform a clean build if API levels don't match
This could happen if for instance we launch an app to a non-running
emulator. On the first launch, the API level isn't known, so 1 is used.
But on a subsequent launch, the API level is known. Even though the
build timestamps match, we still need a rebuild..

Change-Id: I1eff324288ce6f0a089350ceacc8c98d4ad40be7
(cherry picked from commit 1b5d19bc5ce2036df68a2853499036362f169e28)