"Fix Gradle version" quick fix did not work with older versions of Android plugin.
When using older versions of the Android Gradle plugin (e.g. 1.2.0) with
an supported version of Gradle (e.g. 2.10), the IDE displayed the
"Fix Gradle version" quick-fix hyperlink, but when clicked, the hyperlink
did not update the Gradle version.

The reason is that for this version of the Android Gradle plugin, this
error does not come as a "Sync Issue", but as a exception, so the code
that parses the message and abstract the supported version of Gradle was
not being invoked (the code in ProjectSyncMessages.)

The fix is to move the message parsing code to UnsupportedGradleVersionHandler,
so this functionality is available to older versions of the Android Gradle
plugin as well.

Change-Id: Id169b83374304b6c1c9c2ebd6e6d7a7145bd2b42