Allow obtaining device specific properties from not-yet-running AVDs
When an emulator is selected to be the deploy target for a launch, the
gradle build is invoked in parallel while the AVD is launched. Since the
AVD is not yet running during the build, GradleInvokerOptions is not
able to obtain device specific properties necessary for the build.

However, the device properties we need for Gradle are known statically
for AVDs, and can be obtained if the target devices are maintained as
AndroidDevices rather than IDevices. Hence this CL changes the
DeviceFutures class to be based on AndroidDevice rather than IDevice.

Change-Id: I339b9c3d9907afeb89df85d9dbd4aaf243cc01f1
(cherry picked from commit d42c735ef54d41416570893bf4f6c1363e080a3c)