Retrieving sdk handler statically to support changes
Right now we keep a cached version of AndroidSdkHandler in various
places in the sdk manager UI. However, within that UI it's possible to
change the local sdk path, which means a change in the "active"
AndroidSdkHandler instance.
Making the handler be retrived each time rather than keeping instances
around.

Change-Id: I756478625623ea7f40473874b74138703f6f4c82