Test that ResourceFolderRepository tolerates "illegal" files
Test a file with duplicated resource names.
Test a file with an illegal character "*".

This is in preparation for having ResourceFolderRepository
parse plain Files instead of PsiFiles when first loading,
and lazily converting to PsiFiles on edit. We need to
tolerate files that were saved this way to preserve
the existing behavior.

Bug: b/25078759
Change-Id: Ie2156daf709a8e655e3e82e928a86b15efbd2ae3