Don't hotswap when there is no active run session
This fixes the following error:
 IllegalArgumentEx @ LaunchTaskRunner:-1 < AndroidRunState:114

This happens in the following scenario:
 - Run app
 - Kill app from IDE (i.e. terminate session)
 - Make hotswappable change
 - Launch app on device
 - Now press Run

In such a case, we detect that the change is simple and can be
hotswapped, however the hotswap task doesn't know how to create a run
console and process handler. This CL fixes the issue by only allowing
hotswap if there is an existing active run session.

Change-Id: Iff44e84a3750d42ce4218168be9a4fae849dc608