Tweak how multiple devices are handled with instant run
Instant Run doesn't support launching on multiple devices concurrently.
As a result, selection of multiple devices was not allowed in the device
picker. This was unintuitive because as a user, there was no feedback as
to why multi device selection was disabled.

This CL tweaks this behavior by allowing selection of multiple devices
in the device picker, but at launch time, a instant run warning is
displayed indicating that this launch does not use instant run.

Change-Id: I93adb52d9e18d3b21b392604470a97cd4f56efd1
(cherry picked from commit 0c31891721c99bdf808a3e6627d921734b60507a)