Terminate process if the debugger is stopped
This CL re-introduces the earlier CL that kills the process when the
debugger is stopped as required by UX. This was earlier reverted because
of an unexpected interaction with cold swap where killing the process
from the debug process termination listener ended up killing the newly
launched process instead. We guard against this scenario by making sure
that we kill the process only if its pid matches what we originally
connected the debugger to.

Fixes https://code.google.com/p/android/issues/detail?id=195922

Change-Id: Iffd56495b939fbf830a4bf1a272430e73325f6ff