Handle coldswaps being disabled on device API < 21
The Gradle plugin will now return an empty artifact list with a
failing verifier status when there was a verifier failure on a platform
that doesn't support coldswap. This is used to trigger a full build.

Until now, a full build has meant "manifest changed", so the error
messages needed to be tweaked a bit. It also prepends the previous
verifier failure message since otherwise it would be hard to see (it's
shown and then immediately replaced by the "need to do a full build"
bubble.

Change-Id: I8c2bff22fa34d1224a38ffbedc48494c94a46de2