Instant Run: Handle MAIN_SPLIT artifact type
The Gradle plugin now distinguishes between a main APK (used in a non-split
scenario, to be installed with adb install) and a main split APK (to be
installed with adb install-multiple). This removes some hardcoded logic
where we have to guess which install task to be used based on the API level.

Change-Id: Ica9ec7c656df3449fefb63a1b5d9d753f5107a89