Rework the system image selection pane into a tabbed view.
Remove the 2 check boxes for recommended and downloadable images.
Instead show 3 tabs: Recommended, X86 Images, Other Images
which should be a more intuitive UI.
Bug: http://b.android.com/199325

Details: There are now 3 instances of SystemImageList (one for
each tab), the model is shared between them. The coordination
between the tabs are handled by the form ChooseSystemImageStep
which also now holds the refresh button and the status panel.

Documentation Impact:
The System Image selection which is used in 2 cases:
- when creating a new AVD
- when editing the system image for an existing AVD
The selection dialog now has a tab control with 3 tabs:
- Recommended  (the best images to use)
- X86          (the next best images talk to Vince)
- Others       (these images will probably be very slow)
Instead of 2 confusing check boxes.

Change-Id: I34282caa4c388525f81a08c831e347ca5ac3a38b