Use resource repos for ResourceManager#getResourceNames
Caches so that async editor services take less compute
and scanning to process R.type.field names. E.g., in
some large projects with lots of strings and drawables
(~20000), the processing time is now under 10ms where
before it would take over 200ms.

There was some difference in IDs indexed by the old
ResourceManager code and ResourceRepos:

- drawables weren't indexed previously (fixed separately)
- AAR R.txt files seem to include more IDs, coming from
  declare-styleable attributes

We can't quite delete the old code path and the
AndroidValueResourcesIndex (or ID index) because those
are used by the SystemResourceManager to index framework
resources, and there isn't the equivalent yet. It's
also unclear when the "type" queried is not a
ResourceType so we leave the old code as-is.

Bug: b/25078759
Change-Id: Ib606e95fd7d5c0508f4a4b9b30c8458f1010bfaa