Remove RenderSession reference from RenderResult
This refactoring removes RenderSession from RenderResult. Since
RenderSession can be disposed, having a link from RenderResult makes it
hard to know when the RenderSession can actually be disposed.
Most of the uses of RenderResult.getSession() were simply to get the
Result from layoutlib so this CL also copies that into RenderResult to
make it available.

A future CL will rename RenderResult to RenderTaskResult to fix the
naming issue with the new method RenderResult.getRenderResult()

Change-Id: If6b50a9f395621ee620524d92733cbe603a315ea