[UITests] closeAllProjects syncs test runner with AWT thread
When GuiTestRule closes all projects, some cleanup actions are
enqueued for later execution on the AWT thread.

We now flush the AWT queue at this point. Previously, the main
thread would proceeded to execute the next test while the AWT
thread was still executing code on behalf of the previous test,
making debugging and diagnosing harder.

Change-Id: If90ea9f6c1c90c42ff48a9a69f50554940ca120a