Now, sync is executed after a build, if a sync request was issued while the project was being built.
With this CL, if a Gradle sync is requested while the project
is being built, the IDE will wait for the build to finish and
then request a sync.

Cancelling the build will not help, because Gradle does not
offer any guarantees that the build will be stopped at all. It
can only offer a "best effort". Therefore the safest operation
is to let the build finish and then request a project sync
right after that.

Fixes issue
https://code.google.com/p/android/issues/detail?id=200429

Change-Id: Idb19819f63f9c83fec41463b7c2875088b6468a7